fix(worker): cross-compile ennoworker for arm64 and darwin
- fs_posix.go: convert Stat.Nlink (uint32 on arm64) to uint64
- hooks/runner_other.go: match linuxProcessGroup's *syscall.SysProcAttr
  signature on non-linux builds so cmd.SysProcAttr assignment type-checks

+++ ennoworker/internal/prompts/fs_posix.go
@@ -112,7 +112,7 @@ func fstatOwner(fd int) (owned bool, nlink uint64, err error) {
 	if err := unix.Fstat(fd, &stat); err != nil {
 		return false, 0, err
 	}
-	return stat.Uid == uint32(unix.Geteuid()), stat.Nlink, nil
+	return stat.Uid == uint32(unix.Geteuid()), uint64(stat.Nlink), nil
 }
 
 // readBounded reads from fd until EOF or maxBytes+1 bytes. Returns the data
